% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{check_java_version}
\alias{check_java_version}
\title{Check the version of Java}
\usage{
check_java_version(silent = TRUE, startup = TRUE)
}
\arguments{
\item{silent}{Display or not a message}

\item{startup}{Startup message}
}
\description{
Check the version of Java
}
\examples{
check_java_version()
}
