% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R, R/zzz.R
\docType{data}
\name{jd3_utilities}
\alias{jd3_utilities}
\alias{get_java_version}
\alias{current_java_version}
\alias{minimal_java_version}
\title{Java Utility Functions}
\format{
An object of class \code{integer} of length 1.

An object of class \code{numeric} of length 1.
}
\usage{
get_java_version()

current_java_version

minimal_java_version
}
\description{
These functions are used in all JDemetra+ 3.0 packages to easily interact between R and Java objects.
}
\keyword{datasets}
