% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{java-version}
\alias{java-version}
\alias{get_java_version}
\alias{current_java_version}
\alias{minimal_java_version}
\title{Extract the Java installed version}
\format{
An object of class \code{integer} of length 1.

An object of class \code{integer} of length 1.
}
\usage{
get_java_version()

current_java_version

minimal_java_version
}
\value{
\code{get_java_version()} returns the current java installed and usable
version. It's an integer.
\code{minimal_java_version} is the minimal java version accepted currently by
JDemetra+.
}
\description{
Extract the Java installed version
}
\examples{
print(minimal_java_version)
print(get_java_version())

}
\keyword{datasets}
