% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{java-version}
\alias{java-version}
\alias{get_java_version}
\alias{minimal_java_version}
\alias{check_java_version}
\title{Extract the Java installed version}
\usage{
get_java_version()

minimal_java_version

check_java_version(silent = TRUE, startup = TRUE)
}
\arguments{
\item{silent}{Boolean to indicate if a message should be displayed.}

\item{startup}{Boolean to indicate if the function is launch at the startup
of the package.}
}
\value{
\code{get_java_version()} returns the current java installed and usable
version. It's an integer.
\code{minimal_java_version} is the minimal java version accepted currently by
JDemetra+.
\code{check_java_version()} returns \code{TRUE} or \code{FALSE} if the current version of
Java is greater than or equal to the minimum required version.
}
\description{
Extract the Java installed version
}
\examples{
print(minimal_java_version)
print(get_java_version())
check_java_version()

}
