% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{jd3_utilities}
\alias{jd3_utilities}
\alias{get_java_version}
\alias{current_java_version}
\alias{minimal_java_version}
\title{Java Utility Functions}
\format{
An object of class \code{integer} of length 1.

An object of class \code{integer} of length 1.
}
\usage{
get_java_version()

current_java_version

minimal_java_version
}
\value{
\code{get_java_version()} returns the current java installed and usable
version. It's an integer.
}
\description{
Java Utility Functions
}
\examples{
print(minimal_java_version)
print(get_java_version())
}
\keyword{datasets}
