% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{reload_safactories}
\alias{reload_safactories}
\title{Reload all seasonal adjustment factories}
\usage{
reload_safactories()
}
\value{
invisibly \code{NULL}
}
\description{
Reload all seasonal adjustment factories
}
\examples{
\dontshow{if (check_java_version()) withAutoprint(\{ # examplesIf}
reload_safactories()
\dontshow{\}) # examplesIf}
}
